+++ textutil/textutil_seek.c
@@ -2,7 +2,7 @@
 #include <libft/string.h>
 #include <libft/std.h>
 
-int textutil_seek(const char *octets, int olen, const char *pattern, int patlen)
+int textutil_seek(const char *octets, size_t olen, const char *pattern, size_t patlen)
 {
 	if (patlen > 0 && patlen <= olen && ft_strncmp(octets, pattern, patlen) == 0) {
 		return (patlen);
